refactor(data): 👷 objects into data files
