+++ app/work-experience/smartsupp-help/page.tsx
@@ -1,6 +1,6 @@
 import NextPageNavigation from '@/components/NextPageNavigation'
 import { ProjectPageLayoutWrapper } from '@/components/layout/projectPage/ProjectPageLayoutWrapper'
-import { projectsWorkWordPress } from '@/data/projects/work/projectsOverview'
+import { projectsWorkWordPress } from '@/data/projects/work/projectsOverview/workWordpress'
 import { sections } from '@/data/projects/work/smartsuppHelp'
 import { TEXT } from '@/localization/texts_en'
 import { ID, PAGES_URL, URL_WORK_EXPERIENCE } from '@/utils/constants'
