+++ components/homepage/Skills.tsx
@@ -1,5 +1,5 @@
 import DividerWithText from '@/components/shared/DividerWithText'
-import { iconsSkills1, iconsSkills2, iconsSkills3, iconsSkills4 } from '@/data/skills/skills-main'
+import { iconsSkills1, iconsSkills2, iconsSkills3, iconsSkills4 } from '@/data/skills/skillsMain'
 import { ID } from '@/utils/constants'
 import SkillsIconGroup from './SkillsIconGroup'
