+++ data/projects/personal/projectsOverview/index.ts
@@ -1,0 +1,2 @@
+export * from './personalNext'
+export * from './personalReact'
