+++ data/projects/work/projectsOverview/index.ts
@@ -1,0 +1,5 @@
+export * from './workFrontEnd'
+export * from './workLocalization'
+export * from './workQA'
+export * from './workReact'
+export * from './workWordpress'
