+++ data/whatido/skillsInfo.ts
@@ -19,24 +19,3 @@ export const skillsInfo: SkillInfo[] = [
     description: `With ${YEARS.typescript} years of experience in TypeScript, I focus on code safety, and creating robust and scalable codebases. Less errors save both time and\u00A0money.`,
   },
 ]
-
-export const mindsetInfo: SkillInfo[] = [
-  {
-    id: 1,
-    title: 'Continuous learning',
-    description:
-      'I continuously improve my technical skills and stay updated with trends by actively engaging in online communities and\u00A0learning to\u00A0enhance my knowledge and\u00A0expertise.',
-  },
-  {
-    id: 2,
-    title: 'Team player',
-    description:
-      'I enjoy working in teams where I can contribute ideas. My strong communication skills enable me to explain concepts effectively to\u00A0stakeholders and\u00A0team members.',
-  },
-  {
-    id: 3,
-    title: 'Delivering meaningful products',
-    description:
-      'I am passionate about solving real-world problems to help users and customers. Products that streamline processes, and\u00A0increase value of a\u00A0business.',
-  },
-]
